+++ changelogs/fragments/546-constraints.yml
@@ -1,0 +1,2 @@
+minor_changes:
+  - "Support a constraints file that allows to fix dependencies for the ``new-ansible`` and ``prepare`` subcommands (https://github.com/ansible-community/antsibull/pull/546)."
